+++ src/de/omegazirkel/risingworld/DiscordWebHook.java
@@ -75,7 +75,7 @@ import org.json.simple.JSONObject;
  */
 public class DiscordWebHook extends Plugin implements Listener, FileChangeListener {
 
-	public static final String pluginVersion = "0.15.3";
+	public static final String pluginVersion = "0.15.5";
 	public static final String pluginName = "DiscordPlugin";
 	static final String pluginCMD = "dp";