+++ src/de/omegazirkel/risingworld/commands/BroadcastCommand.java
@@ -15,11 +15,12 @@ import net.risingworld.api.Server;
 public class BroadcastCommand implements CommandExecutor {
 
     static final String command = "bc";
+    static final String fullCommand = "/" + command;
     static final String description = "tbd";
 
-    @Command(aliases = { "/"
-            + command }, description = description, async = true, privateMessages = false, usage = "/bc [type] [text]", showInHelpPage = true)
-    public String onCommand(TextChannel channel, Message message, MessageAuthor author, String type, String content) {
+    @Command(aliases = { fullCommand }, description = description, async = true, privateMessages = false, usage = "/bc [type] [text]", showInHelpPage = true)
+    public String onCommand(TextChannel channel, Message message, MessageAuthor author, String cmd, String type,
+            String content) {
         DiscordWebHook plugin = JavaCordBot.pluginInstance;
         Server server = plugin.getServer();
         String lang = plugin.getBotLanguage();