+++ src/de/omegazirkel/risingworld/commands/MakeAdminCommand.java
@@ -16,11 +16,12 @@ import net.risingworld.api.objects.Player;
 public class MakeAdminCommand implements CommandExecutor {
 
     static final String command = "mkadmin";
+    static final String fullCommand = "/" + command;
     static final String description = "tbd";
 
-    @Command(aliases = { "/"
-            + command }, description = description, async = true, privateMessages = false, usage = "/mkadmin [playerName]", showInHelpPage = true)
-    public String onCommand(TextChannel channel, Message message, MessageAuthor author, String playerName) {
+    @Command(aliases = { fullCommand }, description = description, async = true, privateMessages = false, usage = "/mkadmin [playerName]", showInHelpPage = true)
+    public String onCommand(TextChannel channel, Message message, MessageAuthor author, String cmd,
+            String playerName) {
         DiscordWebHook plugin = JavaCordBot.pluginInstance;
         Server server = plugin.getServer();
         String lang = plugin.getBotLanguage();